requirements.txt
aiogram==2.9.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
